Projeto/backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
python-jose
python-multipart
pydantic
pydantic-settings
requests
